domain, view 폴더 추가
